feat: change select func to select class
