+++ lightllm/server/httpserver_for_pd_master/manager.py
@@ -25,6 +25,11 @@ from lightllm.server.metrics.manager import MetricClient
 from lightllm.utils.statics_utils import MovingAverage
 from lightllm.server.httpserver.manager import AsyncQueue
 from lightllm.utils.error_utils import ServerBusyError
+from .pd_selector import (
+    create_selector,
+    PDSelector,
+    MemorySelector,
+)
 
 logger = init_logger(__name__)
 
@@ -35,8 +40,11 @@ class PDManager:
         self.prefill_nodes: List[PD_Client_Obj] = []
         self.decode_nodes: List[PD_Client_Obj] = []
         self.url_to_pd_nodes: Dict[str, PD_Client_Obj] = {}
-        self.select_p_d_node_func = self._get_select_p_d_node_func(args.select_p_d_node_func)
+        self.selector: PDSelector = self._create_selector(args.select_p_d_node_func)
         return
+
+    def _create_selector(self, select_p_d_node_func_name: str) -> PDSelector:
+        return create_selector(select_p_d_node_func_name, self.prefill_nodes, self.decode_nodes)
 
     def register_pd(self, pd_info_json, websocket):
         pd_client = PD_Client_Obj(**pd_info_json)
@@ -51,6 +59,10 @@ class PDManager:
         else:
             assert False
 
+        # 更新选择器的节点列表
+        self.selector.prefill_nodes = self.prefill_nodes
+        self.selector.decode_nodes = self.decode_nodes
+
         logger.info(f"mode: {pd_client.mode} url: {pd_client.client_ip_port} registed")
         return
 
@@ -61,124 +73,22 @@ class PDManager:
         except:
             pass
 
-        # 从内存缓存中删除该节点的数据
-        if hasattr(self, 'node_memory_cache'):
-            self.node_memory_cache.pop(pd_client.client_ip_port, None)
+        # 从内存缓存中删除该节点的数据(如果是MemorySelector)
+        if isinstance(self.selector, MemorySelector):
+            self.selector.remove_node_cache(pd_client.client_ip_port)
 
         self.prefill_nodes = [e for e in self.prefill_nodes if e.client_ip_port != pd_client.client_ip_port]
         self.decode_nodes = [e for e in self.decode_nodes if e.client_ip_port != pd_client.client_ip_port]
+
+        # 更新选择器的节点列表
+        self.selector.prefill_nodes = self.prefill_nodes
+        self.selector.decode_nodes = self.decode_nodes
+
         logger.info(f"mode: {pd_client.mode} url: {pd_client.client_ip_port} removed")
         return
 
-    def _get_select_p_d_node_func(self, select_p_d_node_func_name: str):
-        if select_p_d_node_func_name == "random":
-            return self._random_select_p_d_node
-        elif select_p_d_node_func_name == "round_robin":
-            self.prefill_node_index = 0
-            self.decode_node_index = 0
-            return self._round_robin_select_p_d_node
-        elif select_p_d_node_func_name == "memory":
-            # 内存使用情况缓存
-            self.node_memory_cache: Dict[str, float] = {}  # 节点IP:PORT -> 内存使用率
-            self.memory_monitor_task = None
-            self._start_memory_monitor()
-            return self._memory_select_p_d_node
-        elif select_p_d_node_func_name == "radix":
-            return self._radix_select_p_d_node
-        else:
-            raise ValueError(f"invalid select_p_d_node_func_name: {select_p_d_node_func_name}")
-
-    def _start_memory_monitor(self):
-        """启动内存监控后台任务"""
-        import asyncio
-
-        if self.memory_monitor_task is None or self.memory_monitor_task.done():
-            try:
-                loop = asyncio.get_event_loop()
-                self.memory_monitor_task = loop.create_task(self._memory_monitor_loop())
-                logger.info("Started memory monitoring task")
-            except RuntimeError:
-                logger.warning("No event loop running, memory monitoring will start later")
-
-    async def _memory_monitor_loop(self):
-        """后台内存监控循环，每秒更新一次所有节点的内存使用情况"""
-        import aiohttp
-        import asyncio
-
-        async def get_node_memory_usage(node: PD_Client_Obj) -> tuple:
-            try:
-                node_url = f"http://{node.client_ip_port}/token_load"
-                timeout = aiohttp.ClientTimeout(total=3.0)
-                async with aiohttp.ClientSession(timeout=timeout) as session:
-                    async with session.get(node_url) as response:
-                        if response.status == 200:
-                            data = await response.json()
-                            current_load = data.get("current_load", 0.0)
-                            if isinstance(current_load, list):
-                                current_load = current_load[0] if current_load else 0.0
-                            return node.client_ip_port, float(current_load)
-                        else:
-                            logger.warning(f"Failed to get token_load from {node.client_ip_port}, status: {response.status}")
-                            return node.client_ip_port, float('inf')
-            except Exception as e:
-                logger.warning(f"Error getting memory usage from {node.client_ip_port}: {str(e)}")
-                return node.client_ip_port, float('inf')
-
-        while True:
-            try:
-                all_nodes = self.prefill_nodes + self.decode_nodes
-                tasks = [get_node_memory_usage(node) for node in all_nodes]
-                results = await asyncio.gather(*tasks, return_exceptions=True)
-
-                # 更新缓存
-                for result in results:
-                    if isinstance(result, tuple) and len(result) == 2:
-                        node_key, usage = result
-                        self.node_memory_cache[node_key] = usage
-
-                logger.debug(f"Updated memory cache: {self.node_memory_cache}")
-
-                # 等待1秒后再次检查
-                await asyncio.sleep(1.0)
-
-            except Exception as e:
-                logger.error(f"Error in memory monitor loop: {str(e)}")
-                await asyncio.sleep(1.0)
-
     async def select_p_d_node(self, prompt: Union[str, List[int]], sampling_params: SamplingParams, multimodal_params: MultimodalParams) -> Tuple[PD_Client_Obj, PD_Client_Obj]:
-        assert self.select_p_d_node_func is not None, "select_p_d_node_func is not set"
-        return await self.select_p_d_node_func(prompt, sampling_params, multimodal_params)
-
-    async def _random_select_p_d_node(self, prompt: Union[str, List[int]], sampling_params: SamplingParams, multimodal_params: MultimodalParams) -> Tuple[PD_Client_Obj, PD_Client_Obj]:
-        import random
-
-        p_node = random.choice(self.prefill_nodes)
-        d_node = random.choice(self.decode_nodes)
-        return p_node, d_node
-
-    async def _round_robin_select_p_d_node(self, prompt: Union[str, List[int]], sampling_params: SamplingParams, multimodal_params: MultimodalParams) -> Tuple[PD_Client_Obj, PD_Client_Obj]:
-        p_node = self.prefill_nodes[self.prefill_node_index]
-        d_node = self.decode_nodes[self.decode_node_index]
-        self.prefill_node_index = (self.prefill_node_index + 1) % len(self.prefill_nodes)
-        self.decode_node_index = (self.decode_node_index + 1) % len(self.decode_nodes)
-        return p_node, d_node
-
-    # 哪个节点内存占用最少，就选哪个
-    async def _memory_select_p_d_node(self, prompt: Union[str, List[int]], sampling_params: SamplingParams, multimodal_params: MultimodalParams) -> Tuple[PD_Client_Obj, PD_Client_Obj]:
-        # 获取prefill节点的内存使用情况
-        prefill_usages = [self.node_memory_cache.get(node.client_ip_port, float('inf')) for node in self.prefill_nodes]
-        decode_usages = [self.node_memory_cache.get(node.client_ip_port, float('inf')) for node in self.decode_nodes]
-
-        import random
-        min_prefill_usage = min(prefill_usages)
-        min_decode_usage = min(decode_usages)
-        p_node = self.prefill_nodes[prefill_usages.index(min_prefill_usage)] if min_prefill_usage != float('inf') else random.choice(self.prefill_nodes)
-        d_node = self.decode_nodes[decode_usages.index(min_decode_usage)] if min_decode_usage != float('inf') else random.choice(self.decode_nodes)
-
-        return p_node, d_node
-
-    async def _radix_select_p_d_node(self, prompt: Union[str, List[int]], sampling_params: SamplingParams, multimodal_params: MultimodalParams) -> Tuple[PD_Client_Obj, PD_Client_Obj]:
-        pass
+        return await self.selector.select_p_d_node(prompt, sampling_params, multimodal_params)
 
 class HttpServerManagerForPDMaster:
     def __init__(
@@ -483,8 +393,8 @@ class HttpServerManagerForPDMaster:
         asyncio.create_task(self.timer_log())
 
         # 如果使用memory策略，确保监控任务已启动
-        if hasattr(self.pd_manager, 'memory_monitor_task') and self.args.select_p_d_node_func == "memory":
-            self.pd_manager._start_memory_monitor()
+        if isinstance(self.pd_manager.selector, MemorySelector):
+            self.pd_manager.selector._start_memory_monitor()
 
         use_config_server = self.args.config_server_host and self.args.config_server_port
