+++ lightllm/server/httpserver_for_pd_master/pd_selector/__init__.py
@@ -1,0 +1,29 @@
+from typing import List
+from lightllm.server.httpserver_for_pd_master.manager import PD_Client_Obj
+from .pd_selector import (
+    PDSelector,
+    RandomSelector,
+    RoundRobinSelector,
+    MemorySelector,
+    RadixSelector
+)
+
+__all__ = [
+    "PDSelector",
+    "RandomSelector", 
+    "RoundRobinSelector",
+    "MemorySelector",
+    "RadixSelector"
+]
+
+def create_selector(selector_type: str, prefill_nodes: List[PD_Client_Obj], decode_nodes: List[PD_Client_Obj]) -> PDSelector:
+    if selector_type == "random":
+        return RandomSelector(prefill_nodes, decode_nodes)
+    elif selector_type == "round_robin":
+        return RoundRobinSelector(prefill_nodes, decode_nodes)
+    elif selector_type == "memory":
+        return MemorySelector(prefill_nodes, decode_nodes)
+    elif selector_type == "radix":
+        return RadixSelector(prefill_nodes, decode_nodes)
+    else:
+        raise ValueError(f"Invalid selector type: {selector_type}")
